+++ llm-test-10.yaml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python>=3.12,<4.0
+  - python=3.11
   - pip
   - pip:
       - streamlit==1.45.1
