+++ utils/ui.py
@@ -1,5 +1,7 @@
 # utils/ui.py
 import streamlit as st
+import csv
+
 
 # 메인으로 돌아가는 버튼 
 def render_main_header():
@@ -22,3 +24,12 @@ def render_flag_sub(flag):
             st.success("✅ 정답입니다!")
         else:
             st.error("❌ 틀렸습니다.")
+
+#ctf01 사용하는 scv파일 읽기 함수
+def csv_read_func(file_path):
+    with open(file_path, "r", encoding="utf-8") as f:
+        reader = csv.DictReader(f)
+        lines = []
+        for row in reader:
+            lines.append(f"- 주문자: {row['주문자']} / 상품코드: {row['상품코드']} / 상품명: {row['상품명']} / 가격: {int(row['가격']):,}원")
+        return "\n".join(lines)
